Created new files, parsed the code, to add snake.h
