+++ Snake-AI/Globals.h
@@ -1,0 +1,7 @@
+#pragma once
+
+constexpr int WINDOW_SIZE = 800;
+constexpr int PIXEL_SIZE = 20;
+constexpr int BLOCK_SIZE = WINDOW_SIZE / PIXEL_SIZE;
+constexpr int FPS_LIMIT = 60;
+constexpr float SNAKE_SPEED = 7.f;
