+++ Snake-AI/Snake-AI.vcxproj
@@ -150,6 +150,8 @@
     <ClCompile Include="SnakeGame.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="DrawObjects.h" />
+    <ClInclude Include="Globals.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="SnakeGame.h" />
   </ItemGroup>
