+++ Snake-AI/SnakeGame.h
@@ -4,41 +4,8 @@
 #include <chrono>
 #include <vector>
 
-
-
-constexpr int WINDOW_SIZE = 800;
-constexpr int PIXEL_SIZE = 20;
-constexpr int BLOCK_SIZE = WINDOW_SIZE / PIXEL_SIZE;
-constexpr int FPS_LIMIT = 60;
-constexpr float SNAKE_SPEED = 7.f;
-
-class DrawObject {
-public:
-    virtual void draw(sf::RenderWindow& window) = 0;
-    ~DrawObject() = default;
-};
-
-class Grid : public DrawObject {
-public:
-    Grid(int size) : block(sf::Vector2f(size, size)) {}
-
-    virtual void draw(sf::RenderWindow& window) override {
-        for (int i = 0; i < PIXEL_SIZE; i++) {
-            for (int j = 0; j < PIXEL_SIZE; j++) {
-                block.setPosition(sf::Vector2f(BLOCK_SIZE * j, BLOCK_SIZE * i));
-
-                sf::Color primary(60, 60, 60);
-                sf::Color secondary(120, 120, 120);
-                block.setFillColor(((i + j) % 2) ? primary : secondary);
-
-                window.draw(block);
-            }
-        }
-    }
-private:
-    sf::RectangleShape block;
-
-};
+#include "DrawObjects.h"
+#include "Globals.h"
 
 class SnakeGame {
 public:
@@ -81,7 +48,7 @@ private:
     std::chrono::high_resolution_clock::time_point lastTime, now;
     float fps;
 
-    std::vector<std::unique_ptr<DrawObject>> objects;
+    std::vector<std::unique_ptr<DrawObject>> draw_objects;
 
     std::vector<std::string> Errors;
